Plot 2
Plot 2 shows the measured (actual) load and the load predicted by the
gradient boosting model as well as the error between the actual and
predicted loads. We see that on average our model gives reasonably
accurate predictions. However, there are several problematic time
periods. It is easy to see that many of the points where the error
magnitude is particularly large correspond to holidays: Fourth of July,
Labor Day, Thanksgiving, Christmas, New Year's. Our model could be
accordingly improved simply by incorporating these holidays into the
model as appropriate indicator variables. However, there remains a
problematic period at the end of October and beginning of November 2012.
It looks like in this case the poor performance of our model is related
to some unusual weather conditions. In fact on 27 October, 2012 New York
City (along with other parts of the East Coast) was hit by the Hurricane
Sandy (the second-costliest hurricane in United States history). It is
thus not surprising that our model prediction is not accurate during
that period since the actual power demand was unusually low during such
a severe storm.